+++ Scripts/Signals/LevelSkippedSignal.cs
@@ -2,7 +2,7 @@ namespace TheOneStudio.UITemplate.UITemplate.Signals
 {
     public class LevelSkippedSignal
     {
-        public int Level;
-        public int Time;
+        public int    Level;
+        public double Time;
     }
 }
